+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-ledger, keyring, util, util-crypto 13.5.3
 - extension-dapp 0.61.1
 - extension-dapp 0.60.1
 - react-{identicon, qr}, ui-{keyring, settings}, vue-identicon 3.15.1